Fix compatibility with perl 5.20 and non-dot locales
In locales with something else than a dot as the decimal separator, the
combination of perl >= 5.20 and Gtk2 lead to errors wherever perl tried to
parse version numbers, as in "use 5.8.0" or "use Encode 0.5".  Fix by making
sure we notify perl when the locale might have changed behind its back.